convert filter accordion to use va component
